+++ src/main/java/com/turkcell/loanmodule/business/abstracts/ICreditHistoryService.java
@@ -3,7 +3,6 @@ package com.turkcell.loanmodule.business.abstracts;
 import com.turkcell.loanmodule.entities.concretes.Credit;
 import com.turkcell.loanmodule.entities.concretes.CreditHistory;
 import com.turkcell.loanmodule.entities.concretes.Customer;
-import java.math.BigDecimal;
 import java.time.LocalDate;
 import java.util.List;
 
@@ -11,7 +10,7 @@ public interface ICreditHistoryService {
 
   List<CreditHistory> findAllByCustomer(Customer customer);
 
-  BigDecimal getCreditScoreOfCustomer(Customer customer);
+  Integer getCreditScoreOfCustomer(Customer customer);
 
   CreditHistory save(CreditHistory creditHistory);
   void saveAllCreditHistory(Credit credit) throws Exception;
